requirements.txt
# VegaOCR - Multi-Engine ML-Powered OCR Utility
# Core ML OCR engines
paddlepaddle==2.6.2
paddleocr>=2.7.3
easyocr>=1.7.1
pytesseract>=0.3.10

# Image processing & ML
opencv-python-headless>=4.9.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.22.0
scipy>=1.12.0

# Web backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
websockets>=12.0

# PDF support
PyMuPDF>=1.23.0

# Utilities
aiofiles>=23.2.1
pydantic>=2.5.0
loguru>=0.7.2
